fix(core): frame digest lengths as fixed-width u64
RequestFingerprint::of persists across a host restart and is
re-validated on resume, but every length it fed into the digest —
the attachment count and DigestWriter::frame's own length prefix —
went in as usize::to_le_bytes(). usize is 4 bytes on a 32-bit target
and 8 on a 64-bit one, so the same logical request hashed
differently depending on the platform width. Cast each length to
u64 before framing it so the digest is stable across targets.

+++ crates/mango-external-agents/src/recovery.rs
@@ -45,7 +45,7 @@ impl RequestFingerprint {
         // of file, for no extra distinguishing power — and reaching it at all would mean deriving
         // `Serialize` on the public `Attachment`, handing a host's ids, names and file contents to
         // any serialiser. That is precisely what its redacting `Debug` refuses to do.
-        writer.frame(&request.attachments.len().to_le_bytes());
+        writer.frame(&(request.attachments.len() as u64).to_le_bytes());
         for attachment in &request.attachments {
             writer.frame(attachment.id.as_bytes());
             writer.frame(attachment.name.as_bytes());
@@ -66,7 +66,7 @@ impl DigestWriter {
     /// Feeds one field, prefixed by its length, so a value cannot be split or joined differently
     /// by another arrangement of the same bytes.
     fn frame(&mut self, bytes: &[u8]) {
-        self.0.update(&bytes.len().to_le_bytes());
+        self.0.update(&(bytes.len() as u64).to_le_bytes());
         self.0.update(bytes);
     }
 }
